+++ .github/workflows/build.yml
@@ -5,8 +5,8 @@ jobs:
   build:
     strategy:
       matrix:
-        go-version: [1.11.x, 1.12.x, 1.13.x, 1.14.x, 1.15.x]
-        os: [ubuntu-latest]
+        go-version: [1.12.x, 1.13.x, 1.14.x, 1.15.x]
+        os: [ubuntu-latest, macos-latest]
     runs-on: ${{ matrix.os }}
     env:
       GO111MODULE: "on"
