+++ filesystems_darwin.go
@@ -1,0 +1,22 @@
+// +build darwin
+
+package main
+
+func isLocalFs(m Mount) bool {
+	//FIXME: implement
+	return false
+}
+
+func isFuseFs(m Mount) bool {
+	//FIXME: implement
+	return false
+}
+
+func isNetworkFs(m Mount) bool {
+	//FIXME: implement
+	return false
+}
+
+func isSpecialFs(m Mount) bool {
+	return m.Fstype == "devfs"
+}
